Move BLAST functions to bio.py module
